@@ -1,6 +1,6 @@
 {
   "name": "igo",
-  "version": "18.0.0",
+  "version": "19.0.0-next.27",
   "description": "Infrastructure géomatique ouverte",
   "author": "Communauté IGO",
   "repository": {
@@ -45,49 +45,47 @@
     "release": "node --import tsx scripts/src/release.ts"
   },
   "dependencies": {
-    "@angular/animations": "^18.2.1",
-    "@angular/cdk": "^18.2.1",
-    "@angular/common": "^18.2.1",
-    "@angular/compiler": "^18.2.1",
-    "@angular/core": "^18.2.1",
-    "@angular/forms": "^18.2.1",
-    "@angular/material": "^18.2.1",
-    "@angular/platform-browser": "^18.2.1",
-    "@angular/platform-browser-dynamic": "^18.2.1",
-    "@angular/router": "^18.2.1",
-    "@angular/service-worker": "^18.2.1",
-    "@igo2/auth": "^18.0.0",
-    "@igo2/common": "^18.0.0",
-    "@igo2/context": "^18.0.0",
-    "@igo2/core": "^18.0.0",
-    "@igo2/geo": "^18.0.0",
-    "@igo2/integration": "^18.0.0",
-    "@igo2/utils": "^18.0.0",
+    "@angular/animations": "^19.2.0",
+    "@angular/cdk": "^19.2.0",
+    "@angular/common": "^19.2.0",
+    "@angular/compiler": "^19.2.0",
+    "@angular/core": "^19.2.0",
+    "@angular/forms": "^19.2.0",
+    "@angular/material": "^19.2.0",
+    "@angular/platform-browser": "^19.2.0",
+    "@angular/platform-browser-dynamic": "^19.2.0",
+    "@angular/router": "^19.2.0",
+    "@angular/service-worker": "^19.2.0",
+    "@igo2/auth": "^19.0.0-next.29",
+    "@igo2/common": "^19.0.0-next.29",
+    "@igo2/context": "^19.0.0-next.29",
+    "@igo2/core": "^19.0.0-next.29",
+    "@igo2/geo": "^19.0.0-next.29",
+    "@igo2/integration": "^19.0.0-next.29",
+    "@igo2/utils": "^19.0.0-next.29",
     "core-js": "^3.32.2",
-    "hammerjs": "^2.0.8",
-    "ol": "9.1.0",
+    "ol": "^10.5.0",
     "rxjs": "^7.8.0",
     "tslib": "^2.6.0",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.1",
-    "@angular/cli": "^18.2.1",
-    "@angular/compiler-cli": "^18.2.1",
-    "@cypress/schematic": "^2.5.1",
-    "@stylistic/eslint-plugin": "^2.6.4",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "@types/geojson": "^7946.0.10",
-    "@types/hammerjs": "^2.0.41",
+    "@angular-devkit/build-angular": "^19.2.0",
+    "@angular/cli": "^19.2.0",
+    "@angular/compiler-cli": "^19.2.0",
+    "@cypress/schematic": "^3.0.0",
+    "@stylistic/eslint-plugin": "^4.2.0",
+    "@trivago/prettier-plugin-sort-imports": "^5.2.0",
+    "@types/geojson": "^7946.0.16",
     "@types/jasmine": "^4.3.5",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^20.6.2",
+    "@types/node": "^22.13.0",
     "angular-cli-ghpages": "^2.0.0",
-    "angular-eslint": "18.3.0",
-    "conventional-changelog-cli": "^4.1.0",
-    "cypress": "^13.2.0",
-    "eslint": "^9.9.0",
-    "execa": "^9.3.0",
+    "angular-eslint": "^19.3.0",
+    "conventional-changelog-cli": "^5.0.0",
+    "cypress": "^14.2.0",
+    "eslint": "^9.23.0",
+    "execa": "^9.5.0",
     "http-server": "^14.1.0",
     "jasmine-core": "~5.1.0",
     "jasmine-spec-reporter": "~7.0.0",
@@ -96,9 +94,9 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "prettier": "^3.3.0",
-    "tsx": "^4.17.0",
+    "prettier": "^3.5.0",
+    "tsx": "^4.19.0",
     "typescript": "~5.5.4",
-    "typescript-eslint": "^8.2.0"
+    "typescript-eslint": "^8.29.0"
   }
 }
