@@ -1,7 +1,12 @@
 import { DOCUMENT } from '@angular/common';
 import { provideHttpClient } from '@angular/common/http';
-import { ApplicationConfig } from '@angular/core';
-import { APP_INITIALIZER, Provider, importProvidersFrom } from '@angular/core';
+import {
+  ApplicationConfig,
+  EnvironmentProviders,
+  inject,
+  provideAppInitializer
+} from '@angular/core';
+import { importProvidersFrom } from '@angular/core';
 import { MAT_FORM_FIELD_DEFAULT_OPTIONS } from '@angular/material/form-field';
 import {
   MAT_TOOLTIP_DEFAULT_OPTIONS,
@@ -73,17 +78,19 @@ export const appConfig: ApplicationConfig = {
   ]
 };
 
-function provideTheme(): Provider {
-  return {
-    provide: APP_INITIALIZER,
-    useFactory: (configService: ConfigService, document: Document) => () =>
+function provideTheme(): EnvironmentProviders {
+  return provideAppInitializer(() => {
+    const initializerFn = () => {
+      const configService = inject(ConfigService);
+      const document = inject(DOCUMENT);
       configService.isLoaded$
         .pipe(first((isLoaded) => isLoaded))
         .subscribe(() => {
           const theme = configService.getConfig('theme', DEFAULT_THEME);
           loadTheme(document, theme);
-        }),
-    deps: [ConfigService, DOCUMENT],
-    multi: true
-  };
+        });
+      return;
+    };
+    return initializerFn();
+  });
 }
