@@ -16,6 +16,7 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { HomeButtonComponent } from '@igo2/common/home-button';
 import { IgoInteractiveTourModule } from '@igo2/common/interactive-tour';
 import { PanelComponent } from '@igo2/common/panel';
+import { ResizableBarComponent } from '@igo2/common/resizable-bar';
 import { Tool, Toolbox, ToolboxComponent } from '@igo2/common/tool';
 import { ConfigService } from '@igo2/core/config';
 import { IgoMap, isLayerItem } from '@igo2/geo';
@@ -24,12 +25,13 @@ import { CatalogState, ToolState } from '@igo2/integration';
 import { TranslateModule } from '@ngx-translate/core';
 import { BehaviorSubject, Subscription } from 'rxjs';
 
+import { importAllTools } from './sidenav-import';
+
 @Component({
   selector: 'app-sidenav',
   templateUrl: './sidenav.component.html',
   styleUrls: ['./sidenav.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: true,
   imports: [
     AsyncPipe,
     HomeButtonComponent,
@@ -42,7 +44,8 @@ import { BehaviorSubject, Subscription } from 'rxjs';
     NgIf,
     PanelComponent,
     ToolboxComponent,
-    TranslateModule
+    TranslateModule,
+    ResizableBarComponent
   ]
 })
 export class SidenavComponent implements OnInit, OnDestroy {
@@ -75,6 +78,7 @@ export class SidenavComponent implements OnInit, OnDestroy {
 
   @Output() openedChange = new EventEmitter<boolean>();
   @Output() toolChange = new EventEmitter<Tool>();
+  @Output() widthChange = new EventEmitter<number>();
 
   get toolbox(): Toolbox {
     return this.toolState.toolbox;
@@ -84,7 +88,9 @@ export class SidenavComponent implements OnInit, OnDestroy {
     private toolState: ToolState,
     private configService: ConfigService,
     private catalogState: CatalogState
-  ) {}
+  ) {
+    importAllTools();
+  }
 
   ngOnInit() {
     this.activeTool$$ = this.toolbox.activeTool$.subscribe((tool: Tool) => {
@@ -123,6 +129,10 @@ export class SidenavComponent implements OnInit, OnDestroy {
 
   ngOnDestroy() {
     this.activeTool$$.unsubscribe();
+  }
+
+  handleChange(event: MouseEvent): void {
+    this.widthChange.emit(event.clientX);
   }
 
   onPreviousButtonClick() {
